doc: reference https://github.com/ooni/probe/issues/2381
The https://github.com/ooni/probe/issues/2381 issue advocates for each
experiment _fetching_ its own targets through the correct API.

This repository assumes that we have all the targets in a single unified
structure provided by the check-in v2 API.

However, the code in this repository assumes that targets are opaque
and each experiment could handle different targets.

Consider this sentence from the above-mentioned issue:

> we actually have one single kind of experiment, the one which fetches
> its own input, formatted according to the input type understood by the
> experiment, and processes it accordingly.

It is the most important sentence of the issue. The original focus
was on "fetching its own input". However, rereading the sentence after
a few months, it seems the important concept was actually allowing
each experiment to _handle_ its own targets (we called targets "inputs"
in the issue, because we have not had a chance to spell out the
difference between inputs and targets).

This is the reason why I think the work done in this repository
has helped to explore the design space described in the issue above.