feat: support correct progress-bar progress
When you only run a single nettest within a suite and the nettest
is not the last, you still want the suite to reach 100% at the
end. To implement this, we need a specific message to make sure
that we're actually setting it to 100% at the end of a suite.

The cmd/ooniprobe implementation of this functionality is mostly
wrong and it would need some extra work to be fixed.