feat: redesign the view
The interpreter now updates view parameters and the view
periodically refreshes and writes the stdout.

We chain a possibly do-nothing log collector to the view
itself, such that we can write a log file.

There's no real progress bar anymore, which we should maybe
consider of adding again to the mix.