refactor: rename instruction to command
It seems the "command" term is a better term to use considering
that we're dealing with a script. We can say that a script contains
zero or more commands that the engine should execute. This sounds
better than saying that a script contains instructions because
usually we use "instructions" when discussing assembly code.