feat: better composable-mini-nettests implementation (#2)
This diff improves upon #1 to provide a better implementation of mini
nettests. The reference issue is
https://github.com/ooni/probe/issues/2494.

Here's what changed since the previous prototype:

1. *separate layers more clearly*: now we have a public package API that
generates the lisp-like AST. Then, we have a simpler "compiler" than
interprets the AST and uses a template for each existing function to
transform it into a proper function from any to any.

2. *simpler monadic values routing*: we type switch over `any` instead
of having an explicit monad.

3. *better support for generating top-level keys*: now we have a `Skip`
type that a function that generates top-level keys can return once it
has processed some values so we don't trigger subsequent analysis steps
😅

4. *better support for exceptions*: rather than panicking, we return an
`Exception` type with a reason.

5. *better support for options*: options are parsed as normal functions
that throw an `Exception` but implement an interface that allows the
proper consumer to use the options to configure execution (when options
are used in the wrong place, they would therefore cause an `Exception`
to be propagated)

6. *more pragmatic support for observations*: since we have a runtime,
we append the observations to a runtime list in locked fashion, rathern
than moving them around, which simplifies the code *a lot*

7. *better rules for receiving and returning "nothing"*: we expect
`Void` in input and we return `Skip` in output

8. *better composability*: now each `Function` transforms its input just
once rather than multiple times, so it is easier to reason about what
transformations is applied by each step

9. *better error messages*: we perform most checks when we compile the
lisp-like AST rather than when executing functions, and we added better
code to clearly report what went wrong

10. *reasonable homoiconicity between the internal DSL and lisp-like
AST*: the main difference in this department is that we need to handle
integer types specially when loading from JSON because they're loaded as
`float64` by default.