fix: let the identity behave as the do-nothing option
This change is fundamental to ensuring we can provide conditional
support for options implemented by newer probes only without breaking
existing probes. We'll use "try_compose" (or, even better, a version
of "try_compose" renamed to "if_function_exists"). If the function does
not exist, we return the identity. So, if the identity behaves like
the do-nothing option, we can use this mechanism for options.