fix: HTTP should return a response
If we do not see a response, we cannot use the response to
follow redirects and other goodies. This is not something we
plan on doing for IM tests, but we may want to use unlang's
runtime for implementing Web Connectivity, where we would
need to have this functionality ready.