feat: refactor unruntime to have explicit typing
This diff commits work in progress where we've refactored unruntime to
use explicit types, which makes the new code look similar to the original
dslx code except that now we have complete composability.

I realized I could possibly need pure types and the possibility of
swapping out data collection and use a fake "trace" because this
functionality may be handy for rewriting the TH.

However, it's still not 100% clear to me how we could modify the unruntime
such that under the hood it uses types. But, assuming that would be
possible, then we could serve three use cases:

1. typing, no tracing, just the network behavior (oohelperd)

2. typing, tracing and network behavior (maybe web connectivity v0.6?)

3. no typing, tracing and network behavior (richer input?)