feat: internal and external DSL for writing nettests (#7)
This diff improves the DSL for writing nettests introduced in
https://github.com/ooni/2023-05-richer-input/pull/5 such that we a
single homoiconic representation of the DSL. The internal DSL, written
in pure Go, could either perform a measurement directly or could be
exported to a JSON-based AST. The external DSL derives from the
JSON-serialized AST and runs the same operations as the internal one,
except that type checking happens at runtime.

Because the internal DSL is in pure Go, we don't need to write custom
code for type checking. At the same time, we have now lost the
possibility of having a polymorphic HTTP step and we need to cast a TCP,
TLS, or QUIC connection to become an HTTP connection by using the proper
operation.

All the other functionality should be compatible with
https://github.com/ooni/2023-05-richer-input/pull/5. Obviously, I needed
to change the JSON representation to accommodate for minor changes, so
the previously serialized JSONs should not be working anymore. This is
fine, since we're still in the experimentation stage.

To be more confident that dynamic type checking works, this PR added
comprehensive [netem](https://github.com/ooni/netem) based QA tests that
exercise the most common type conversions happening on success or error.

In terms of terminology, I also chose to replace "func" with "stage" (or
"stage_name") since it seems more selective to deal with the composition
of pipeline stages than with the composition of functions. (The pipeline
stages are functions but, by calling them pipeline stages, we further
clarify what we're dealing with.)

This work is part of https://github.com/ooni/probe/issues/2494 and
https://github.com/ooni/ooni.org/issues/1296. The latter issue is
related to this PR because we also included support for throttling to
the DSL.