feat(dsl): do not include HTTP body by default
By making it explicit that we want to include the HTTP body into a
measurement we highlight that including the body should only be done
when the actual content of the body matters. It makes sense, for
example, for Web Connectivity but it does not make sense to do so
in many other contexts.

Noticed when working on https://github.com/ooni/probe/issues/2502